+++ src/xtgeo/surface/_regsurf_import.py
@@ -3,7 +3,9 @@
 
 import numpy as np
 import numpy.ma as ma
+from struct import unpack
 
+from xtgeo.common.constants import UNDEF_MAP_IRAPB, UNDEF_MAP_IRAPA
 import xtgeo
 import xtgeo.cxtgeo._cxtgeo as _cxtgeo  # pylint: disable=no-name-in-module
 from xtgeo.common import XTGeoDialog
@@ -13,7 +15,7 @@ xtg = XTGeoDialog()
 logger = xtg.functionlogger(__name__)
 
 
-def import_irap_binary(self, mfile, values=True):
+def import_irap_binary(self, mfile, values=True, engine="cxtgeo"):
     """Import Irap binary format.
 
     Args:
@@ -25,6 +27,80 @@ def import_irap_binary(self, mfile, values=True):
         RuntimeError: Problem....
     """
 
+    if mfile.memstream is True or engine == "python":
+        _import_irap_binary_purepy(self, mfile)
+    else:
+        _import_irap_binary(self, mfile, values=values)
+
+
+def _import_irap_binary_purepy(self, mfile, values=True):
+    """Using pure python, better for memorymapping/threading"""
+
+    # Borrowed some code from https://github.com/equinor/grequi/../fmt_irapbin.py
+
+    logger.info("Enter function %s", __name__)
+
+    if mfile.memstream:
+        mfile.file.seek(0)
+        buf = mfile.file.read()
+    else:
+        with open(mfile.file, "rb") as fhandle:
+            buf = fhandle.read()
+
+    # unpack header with big-endian format string
+    hed = unpack(">3i6f3i3f10i", buf[:100])
+
+    self._nrow = hed[2]
+    self._xori = hed[3]
+    self._yori = hed[5]
+    self._xinc = hed[7]
+    self._yinc = hed[8]
+    self._ncol = hed[11]
+    self._rotation = hed[12]
+
+    self._yflip = 1
+    if self._yinc < 0.0:
+        self._yinc *= -1
+        self._yflip = -1
+
+    if not values:
+        self._isloaded = False
+        self._values = None
+        return
+
+    # Values: traverse through data blocks
+    stv = 100  # Starting byte
+    datav = []
+
+    while True:
+        # start block integer - number of bytes of floats in following block
+        blockv = unpack(">i", buf[stv : stv + 4])[0]
+        stv += 4
+        # floats
+        datav.append(
+            np.array(unpack(">" + str(int(blockv / 4)) + "f", buf[stv : blockv + stv]))
+        )
+        stv += blockv
+        # end block integer not needed really
+        _ = unpack(">i", buf[stv : stv + 4])[0]
+        stv += 4
+        if stv == len(buf):
+            break
+
+    values = np.hstack(datav)
+    values = np.reshape(values, (self._ncol, self._nrow), order="F")
+    values = np.array(values, order="C")
+    values = np.ma.masked_greater_equal(values, UNDEF_MAP_IRAPB)
+    self._values = np.ma.masked_invalid(values)
+
+    self._ilines = np.array(range(1, self.ncol + 1), dtype=np.int32)
+    self._xlines = np.array(range(1, self.nrow + 1), dtype=np.int32)
+
+    del buf
+
+
+def _import_irap_binary(self, mfile, values=True):
+
     logger.info("Enter function %s", __name__)
 
     cfhandle = mfile.get_cfhandle()
@@ -34,7 +110,7 @@ def import_irap_binary(self, mfile, values=True):
         ier,
         self._ncol,
         self._nrow,
-        _ndef,
+        _,
         self._xori,
         self._yori,
         self._xinc,
@@ -51,11 +127,6 @@ def import_irap_binary(self, mfile, values=True):
     if self._yinc < 0.0:
         self._yinc *= -1
         self._yflip = -1
-
-    self._filesrc = mfile.name
-
-    self._ilines = np.array(range(1, self._ncol + 1), dtype=np.int32)
-    self._xlines = np.array(range(1, self._nrow + 1), dtype=np.int32)
 
     # lazy loading, not reading the arrays
     if not values:
@@ -82,6 +153,9 @@ def import_irap_binary(self, mfile, values=True):
     self._isloaded = True
     self.values = val
 
+    self._ilines = np.array(range(1, self._ncol + 1), dtype=np.int32)
+    self._xlines = np.array(range(1, self._nrow + 1), dtype=np.int32)
+
     mfile.cfclose()
 
 
@@ -102,55 +176,6 @@ def import_irap_ascii(self, mfile, engine="cxtgeo"):
         _import_irap_ascii_purepy(self, mfile)
     else:
         _import_irap_ascii(self, mfile)
-
-
-def _import_irap_ascii(self, mfile):
-    """Import Irap ascii format via C routines (fast, but less suited for bytesio)"""
-
-    logger.debug("Enter function...")
-    cfhandle = mfile.get_cfhandle()
-
-    # read with mode 0, scan to get mx my
-    xlist = _cxtgeo.surf_import_irap_ascii(cfhandle, 0, 1, 0)
-
-    nvn = xlist[1] * xlist[2]  # mx * my
-    xlist = _cxtgeo.surf_import_irap_ascii(cfhandle, 1, nvn, 0)
-
-    ier, ncol, nrow, _ndef, xori, yori, xinc, yinc, rot, val = xlist
-
-    if ier != 0:
-        mfile.cfclose()
-        raise RuntimeError("Problem in {}, code {}".format(__name__, ier))
-
-    val = np.reshape(val, (ncol, nrow), order="C")
-
-    val = ma.masked_greater(val, xtgeo.UNDEF_LIMIT)
-
-    if np.isnan(val).any():
-        logger.info("NaN values are found, will mask...")
-        val = ma.masked_invalid(val)
-
-    yflip = 1
-    if yinc < 0.0:
-        yinc = yinc * -1
-        yflip = -1
-
-    self._ncol = ncol
-    self._nrow = nrow
-    self._xori = xori
-    self._yori = yori
-    self._xinc = xinc
-    self._yinc = yinc
-    self._yflip = yflip
-    self._rotation = rot
-    self._filesrc = mfile.name
-
-    self.values = val
-
-    self._ilines = np.array(range(1, ncol + 1), dtype=np.int32)
-    self._xlines = np.array(range(1, nrow + 1), dtype=np.int32)
-
-    mfile.cfclose()
 
 
 def _import_irap_ascii_purepy(self, mfile):
@@ -177,7 +202,7 @@ def _import_irap_ascii_purepy(self, mfile):
     values = np.array(buf[19:]).astype(np.float64)
     values = np.reshape(values, (self._ncol, self._nrow), order="F")
     values = np.array(values, order="C")
-    self.values = np.ma.masked_equal(values, 9999900.0000)
+    self.values = np.ma.masked_greater_equal(values, UNDEF_MAP_IRAPA)
 
     self._yflip = 1
     if self._yinc < 0.0:
@@ -186,9 +211,56 @@ def _import_irap_ascii_purepy(self, mfile):
 
     self._ilines = np.array(range(1, self.ncol + 1), dtype=np.int32)
     self._xlines = np.array(range(1, self.nrow + 1), dtype=np.int32)
-    self._filesrc = mfile.name
 
     del buf
+
+
+def _import_irap_ascii(self, mfile):
+    """Import Irap ascii format via C routines (fast, but less suited for bytesio)"""
+
+    logger.debug("Enter function...")
+    cfhandle = mfile.get_cfhandle()
+
+    # read with mode 0, scan to get mx my
+    xlist = _cxtgeo.surf_import_irap_ascii(cfhandle, 0, 1, 0)
+
+    nvn = xlist[1] * xlist[2]  # mx * my
+    xlist = _cxtgeo.surf_import_irap_ascii(cfhandle, 1, nvn, 0)
+
+    ier, ncol, nrow, _, xori, yori, xinc, yinc, rot, val = xlist
+
+    if ier != 0:
+        mfile.cfclose()
+        raise RuntimeError("Problem in {}, code {}".format(__name__, ier))
+
+    val = np.reshape(val, (ncol, nrow), order="C")
+
+    val = ma.masked_greater(val, xtgeo.UNDEF_LIMIT)
+
+    if np.isnan(val).any():
+        logger.info("NaN values are found, will mask...")
+        val = ma.masked_invalid(val)
+
+    yflip = 1
+    if yinc < 0.0:
+        yinc = yinc * -1
+        yflip = -1
+
+    self._ncol = ncol
+    self._nrow = nrow
+    self._xori = xori
+    self._yori = yori
+    self._xinc = xinc
+    self._yinc = yinc
+    self._yflip = yflip
+    self._rotation = rot
+
+    self.values = val
+
+    self._ilines = np.array(range(1, ncol + 1), dtype=np.int32)
+    self._xlines = np.array(range(1, nrow + 1), dtype=np.int32)
+
+    mfile.cfclose()
 
 
 def import_ijxyz_ascii(self, mfile):  # pylint: disable=too-many-locals
@@ -230,7 +302,6 @@ def import_ijxyz_ascii(self, mfile):  # pylint: disable=too-many-locals
     self._nrow = nrow
     self._rotation = rot
     self._yflip = yflip
-    self._filesrc = mfile
 
     self.values = val.reshape((self._ncol, self._nrow))
 
@@ -267,7 +338,6 @@ def import_ijxyz_ascii_tmpl(self, mfile, template):
     self._rotation = template.rotation
     self._yflip = template.yflip
     self.values = val.reshape((self._ncol, self._nrow))
-    self._filesrc = mfile
 
     self._ilines = template._ilines.copy()
     self._xlines = template._xlines.copy()
@@ -286,7 +356,10 @@ def import_petromod_binary(self, mfile, values=True):
     dsc, _ = _cxtgeo.surf_import_petromod_bin(cfhandle, 0, 0.0, 0, 0, 0)
 
     fields = dsc.split(",")
-    print(fields)
+
+    rota_xori = 0
+    rota_yori = 0
+    undef = 999999.0
 
     for field in fields:
         key, value = field.split("=")
@@ -325,6 +398,106 @@ def import_petromod_binary(self, mfile, values=True):
     values = values.reshape(self._ncol, self._nrow)
 
     self.values = values
-    self.filesrc = mfile
+
+    self._ilines = np.array(range(1, self.ncol + 1), dtype=np.int32)
+    self._xlines = np.array(range(1, self.nrow + 1), dtype=np.int32)
 
     mfile.cfclose()
+
+
+def import_zmap_ascii(self, mfile, values=True):
+    """
+    Importing ZMAP + ascii files, in pure python only
+
+    See also:
+    https://mycarta.wordpress.com/2019/03/23/working-with-zmap-grid-files-in-python/
+    https://blog.nitorinfotech.com/what-is-zmap-plus-file-format/
+
+    """
+    if not mfile.memstream:
+        logger.info("Reading zmap+ from %s", mfile.file)
+    else:
+        logger.info("Reading zmap+ from memory stream")
+
+    count = 0
+    buffer = ""
+    self._values = None
+
+    if mfile.memstream:
+        mfile.file.seek(0)
+        buf = mfile.file.read()
+        buf = buf.decode().split("\n")
+        for line in buf:
+            buffer += line
+            count += 1
+            if line.startswith("@\n"):
+                break
+    else:
+        with open(mfile.file, "r", encoding="utf-8") as fhandle:
+            for line in fhandle:
+                buffer += line
+                count += 1
+                if line.startswith("@\n"):
+                    break
+
+    # interpret header
+    correctformat = False
+    header = []
+    lines = buffer.split("\n")
+    for line in lines:
+        if line.startswith("!"):
+            continue
+        elif line.startswith("@ GRIDFILE, GRID"):
+            correctformat = True
+        elif line.startswith("@\n"):
+            break
+        else:
+            hdr = [xxx.strip() for xxx in line.split(",")]
+            header.extend(hdr)
+
+    if not correctformat:
+        raise ValueError("Input file does not seem to be a correct zmap file")
+    print(header)
+
+    self._ncol = int(header[6])
+    self._nrow = int(header[5])
+    self._xori = float(header[7])
+    self._yori = float(header[9])
+    self._xinc = float(header[8]) - float(header[7]) / (self._ncol + 1)
+    self._yinc = float(header[10]) - float(header[9]) / (self._nrow + 1)
+
+    if header[2]:
+        undef = float(header[2])  # user defined undef
+    else:
+        undef = float(header[1])
+
+    logger.info("UNDEF value is %s", undef)
+    print(self._xinc)
+
+    if values is False:
+        self.isloaded = False
+        return
+
+    if mfile.memstream:
+        mfile.file.seek(0)
+        buf = mfile.file.read()
+        buf = buf.decode().split()
+        indexes = [i for i, x in enumerate(buf) if x == "@"]
+        buf = buf[indexes[-1] + 1 :]
+    else:
+        with open(mfile.file, "r", encoding="utf-8") as fhx:
+            buf = fhx.read()
+
+            buf = buf.split()
+            indexes = [i for i, x in enumerate(buf) if x == "@"]
+            buf = buf[indexes[-1] + 1 :]
+
+    values = np.array(buf, dtype=np.float64)
+    values = np.reshape(values, (self._ncol, self._nrow), order="C")
+    values = np.flip(values, axis=1)
+    self._values = np.ma.masked_equal(values, undef)
+
+    self._ilines = np.array(range(1, self.ncol + 1), dtype=np.int32)
+    self._xlines = np.array(range(1, self.nrow + 1), dtype=np.int32)
+
+    logger.info("Reading zmap+... done")